server/requirements.txt
alabaster==0.7.12
alembic==1.2.1
Babel==2.7.0
blinker==1.4
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dialogflow==0.6.0
docutils==0.14
Flask==1.0.3
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
google-api-core==1.13.0
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
packaging==19.0
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyfcm==1.4.7
Pygments==2.4.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.3
urllib3==1.25.3
Werkzeug==0.15.4
